+++ src/security-architecture/service-registration.js
@@ -1,5 +1,6 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { AuthorizationModelDescriptor } from './authorization/authorization-model-descriptor.js';
+import { CryptographyDataProtectionDescriptor } from './cryptography/cryptography-data-protection-descriptor.js';
 import { IdentityAuthenticationDescriptor } from './identity/identity-authentication-descriptor.js';
 import { SecurityOverviewDescriptor } from './overview/security-overview-descriptor.js';
 import { SecretsKeyManagementDescriptor } from './secrets/secrets-key-management-descriptor.js';
@@ -11,6 +12,7 @@ export function addSecurityArchitecture(services) {
   }
 
   services.registerSingleton('AuthorizationModelDescriptor', () => new AuthorizationModelDescriptor());
+  services.registerSingleton('CryptographyDataProtectionDescriptor', () => new CryptographyDataProtectionDescriptor());
   services.registerSingleton('IdentityAuthenticationDescriptor', () => new IdentityAuthenticationDescriptor());
   services.registerSingleton('SecurityOverviewDescriptor', () => new SecurityOverviewDescriptor());
   services.registerSingleton('SecretsKeyManagementDescriptor', () => new SecretsKeyManagementDescriptor());